improve performance for part 1

+++ src/Math/Constants.h
@@ -7,4 +7,4 @@
 #define F_PI ((float)(M_PI))
 #endif
 
-const double CONST_GRAVITY = 9.81; // gravity in [m/s^2]
+const float CONST_GRAVITY = 9.81f; // gravity in [m/s^2]
